Moved 'defaults' to the example folder so they aren't forced.
